+++ src/main.js
@@ -1,16 +1,16 @@
 import { render } from './framework/render.js';
 import FilterView from './view/filter-view.js';
-import TripPlanPresenter from './presenter/plan-presenter.js';
+import PlanPresenter from './presenter/plan-presenter.js';
 import InfoPresenter from './presenter/info-presenter.js';
-import TripEventsModel from './model/trip-events-model.js';
+import EventsModel from './model/events-model.js';
 import { generateFilter } from './mock/filter.js';
 
 const tripMainContainer = document.querySelector('.trip-main');
 const filtersContainer = tripMainContainer.querySelector('.trip-controls__filters');
 const tripPlanContainer = document.querySelector('.trip-events');
-const tripEventsModel = new TripEventsModel();
+const tripEventsModel = new EventsModel();
 
-const tripPlanPresenter = new TripPlanPresenter({
+const tripPlanPresenter = new PlanPresenter({
   tripPlanContainer: tripPlanContainer,
   tripEventsModel,
 });
